requirements.txt
# Core
numpy>=1.26
pyyaml>=6.0
python-dotenv>=1.0

# Embeddings
sentence-transformers==2.6.1

# Vector Store
faiss-cpu==1.8.0

# LLM
transformers==4.41.2
torch>=2.0
